refactor: make the process reaction data to util
